Moved leauthaud model into its own module
